+++ src/main/java/com/renomad/minum/utils/LRUCache.java
@@ -1,7 +1,6 @@
 package com.renomad.minum.utils;
 
 import java.io.Serial;
-import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.Map;