+++ src/main/java/com/renomad/minum/web/Request.java
@@ -57,7 +57,7 @@ public record Request(Headers headers, RequestLine requestLine, Body body,
                       String remoteRequester) {
 
     public static Request EMPTY(Context context) {
-        return new Request(new Headers(List.of(), context), RequestLine.EMPTY(context), Body.EMPTY(context), "");
+        return new Request(new Headers(List.of(), context), RequestLine.EMPTY(context), Body.EMPTY, "");
     }
 
 }