+++ src/test/java/com/renomad/minum/security/TheBrigTests.java
@@ -30,4 +30,26 @@ public class TheBrigTests {
         assertFalse(b.isInJail("1.2.3.4_too_freq_downloads"));
         b.stop();
     }
+
+    /**
+     * If an inmate was in the brig already and they transgressed again,
+     * simply put them back in with a duration starting from the moment
+     * of the new transgression.
+     */
+    @Test
+    public void test_TheBrig_ExistingInmate() {
+        var b = new TheBrig(10, context);
+        b.initialize();
+        b.sendToJail("1.2.3.4_too_freq_downloads", 20);
+        assertTrue(b.isInJail("1.2.3.4_too_freq_downloads"));
+        MyThread.sleep(10);
+        b.sendToJail("1.2.3.4_too_freq_downloads", 20);
+        MyThread.sleep(10);
+        b.sendToJail("1.2.3.4_too_freq_downloads", 20);
+        MyThread.sleep(10);
+        b.sendToJail("1.2.3.4_too_freq_downloads", 20);
+        MyThread.sleep(70);
+        assertFalse(b.isInJail("1.2.3.4_too_freq_downloads"));
+        b.stop();
+    }
 }